in mj828.c/.h: - implemented _DisplayBatteryVoltage(), which lights up battery LEDs according to voltage read out by ADC at Vbat pin
in mj828_led.c:
- corrected a bug: return statement would leave a charlieplexed LED on

+++ mj8x8/Core/Inc/mj828/mj828_adc.c
@@ -28,7 +28,7 @@ static void _ReadChannels(void)
 }
 
 // returns value stored at index i
-static inline uint8_t _GetVal(const uint8_t i)
+static inline uint16_t _GetVal(const uint8_t i)
 {
 	return __ADC.__channel[i];
 }
